+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project will adhere to [Semantic Versioning](https://semver.org/spec/v2
 
 ## [Unreleased]
 
+### Added
+
+- rotation angle for Rect
+- rotate method for Point
+
 ## [0.0.6] - 2023-05-30
 
 ### Added